requirements.txt
flask
flask-cors
torch>=1.7.0
ultralytics
vietocr
Pillow
numpy
requests
